Perform case-insensitive nick comparisons.
We consistently display our nick using its original characters, but
whenever we compare it with message text, we ignore case.  This gives
us a consistent user experience.

Also, the automatic nick-to-mention translation (used when a mention
name is not explicitly configured) has been improved to strip all non-
alphanumeric characters, not just spaces.  This is consistent with
Hipchat's own automatic mention-name-generating rules.